Separate entities by service

+++ src/entity/controllers.py
@@ -1,29 +1,35 @@
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, PermissionDenied
 
 from .models import Entity
 
 
-def create_entity(validated_data):
-    return Entity.objects.create(**validated_data)
+def create_entity(user, validated_data):
+    return Entity.objects.create(user=user, **validated_data)
 
 
-def update_entity(entity, validated_data):
+def update_entity(user, entity, validated_data):
+    if entity.user != user:
+        raise PermissionDenied("You do not have permission to update this entity.")
+
     for attr, value in validated_data.items():
         setattr(entity, attr, value)
     entity.save()
     return entity
 
 
-def delete_entity(entity):
+def delete_entity(user, entity):
+    if entity.user != user:
+        raise PermissionDenied("You do not have permission to delete this entity.")
     entity.delete()
 
 
-def get_entity_by_id(entity_id):
+def get_entity_by_id(user, entity_id):
     try:
-        return Entity.objects.get(id=entity_id)
+        entity = Entity.objects.get(id=entity_id, user=user)
+        return entity
     except Entity.DoesNotExist:
         raise NotFound("Entity not found")
 
 
-def get_entities():
-    return Entity.objects.all()
+def get_entities(user):
+    return Entity.objects.filter(user=user)
